moves base localfx to base.py
