+++ openvair/modules/volume/domain/base.py
@@ -134,7 +134,54 @@ class BaseVolume(metaclass=abc.ABCMeta):
             return {}
 
 
-class RemoteFSVolume(BaseVolume):
+class BaseLocalFSVolume(BaseVolume):
+    """Base class for local filesystem volumes.
+
+    This class provides a base implementation for managing volumes stored on
+    local filesystems.
+    """
+
+    def __init__(self, *args: Any, **kwargs: Any) -> None:  # noqa: ANN401 # TODO need to parameterize the arguments correctly, in accordance with static typing
+        """Initialize a BaseLocalFSVolume instance."""
+        super(BaseLocalFSVolume, self).__init__(*args, **kwargs)
+
+    def create(self) -> Dict:
+        """Create a new volume.
+
+        Raises:
+            NotImplementedError: Must be implemented by subclasses.
+        """
+        raise NotImplementedError
+
+    def delete(self) -> Dict:
+        """Delete an existing volume.
+
+        Raises:
+            NotImplementedError: Must be implemented by subclasses.
+        """
+        raise NotImplementedError
+
+    def extend(self, new_size: str) -> Dict:
+        """Extend an existing volume to the given size.
+
+        Args:
+            new_size (int): The new size for the volume.
+
+        Raises:
+            NotImplementedError: Must be implemented by subclasses.
+        """
+        raise NotImplementedError
+
+    def attach_volume_info(self) -> Dict:
+        """Get information about an existing volume.
+
+        Raises:
+            NotImplementedError: Must be implemented by subclasses.
+        """
+        raise NotImplementedError
+
+
+class BaseRemoteFSVolume(BaseVolume):
     """Abstract base class for RemoteFS volume operations.
 
     This class provides an interface for managing volumes stored on remote
@@ -144,7 +191,7 @@ class RemoteFSVolume(BaseVolume):
     def __init__(self, *args: Any, **kwargs: Any) -> None:  # noqa: ANN401 # TODO need to parameterize the arguments correctly, in accordance with static typing
         """Initialize a RemoteFSVolume instance."""
         LOG.info('Initialized RemoteFSVolume')
-        super(RemoteFSVolume, self).__init__(*args, **kwargs)
+        super(BaseRemoteFSVolume, self).__init__(*args, **kwargs)
 
     def create(self) -> Dict:
         """Create a new volume.
