+++ openvair/modules/volume/domain/physical_fs/localfs.py
@@ -15,56 +15,9 @@ from openvair.libs.log import get_logger
 from openvair.libs.cli.models import ExecuteParams
 from openvair.libs.cli.executor import execute
 from openvair.libs.cli.exceptions import ExecuteError
-from openvair.modules.volume.domain.base import BaseVolume
+from openvair.modules.volume.domain.base import BaseLocalFSVolume
 
 LOG = get_logger(__name__)
-
-
-class BaseLocalFSVolume(BaseVolume):
-    """Base class for local filesystem volumes.
-
-    This class provides a base implementation for managing volumes stored on
-    local filesystems.
-    """
-
-    def __init__(self, *args: Any, **kwargs: Any) -> None:  # noqa: ANN401 # TODO need to parameterize the arguments correctly, in accordance with static typing
-        """Initialize a BaseLocalFSVolume instance."""
-        super(BaseLocalFSVolume, self).__init__(*args, **kwargs)
-
-    def create(self) -> Dict:
-        """Create a new volume.
-
-        Raises:
-            NotImplementedError: Must be implemented by subclasses.
-        """
-        raise NotImplementedError
-
-    def delete(self) -> Dict:
-        """Delete an existing volume.
-
-        Raises:
-            NotImplementedError: Must be implemented by subclasses.
-        """
-        raise NotImplementedError
-
-    def extend(self, new_size: str) -> Dict:
-        """Extend an existing volume to the given size.
-
-        Args:
-            new_size (int): The new size for the volume.
-
-        Raises:
-            NotImplementedError: Must be implemented by subclasses.
-        """
-        raise NotImplementedError
-
-    def attach_volume_info(self) -> Dict:
-        """Get information about an existing volume.
-
-        Raises:
-            NotImplementedError: Must be implemented by subclasses.
-        """
-        raise NotImplementedError
 
 
 class LocalFSVolume(BaseLocalFSVolume):
@@ -106,7 +59,7 @@ class LocalFSVolume(BaseLocalFSVolume):
                     run_as_root=self._execute_as_root,
                     shell=True,
                     raise_on_error=True,
-                )
+                ),
             )
         except (ExecuteError, OSError) as err:
             LOG.error(f'Error while creating volume: {err!s}')
@@ -131,7 +84,7 @@ class LocalFSVolume(BaseLocalFSVolume):
                     run_as_root=self._execute_as_root,
                     shell=True,
                     raise_on_error=True,
-                )
+                ),
             )
         except (ExecuteError, OSError) as err:
             LOG.error(f'Error while deleting volume: {err!s}')
@@ -164,7 +117,7 @@ class LocalFSVolume(BaseLocalFSVolume):
                     run_as_root=self._execute_as_root,
                     shell=True,
                     raise_on_error=True,
-                )
+                ),
             )
         except (ExecuteError, OSError) as err:
             LOG.error(f'Error while extending volume: {err!s}')
